Plotting of kernel with tests
